getPartsInfo: Simplify fetching parts info.
This change also avoids making an unecessary listParts call for
a new multipart upload session.
